pyproject.toml
[project]
name = "prospecting-agent"
version = "0.1.0"
description = "An agentic prospecting tool built on AWS Strands SDK"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_backend"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
